fixed typo in items.
Also, next is the event list. I'm thinking it sets up the whole fight which takes care of duration of effects. It'll just recalculate each tick which should be ok. I'll probably keep it to full ticks, not half. Anyway, there'll be a top level funct called flush which clears what's in the action stack then applies/sec effects like gold and hpregen. it'll finish when one is dead or the stack is empty.

Also, need to add bit for max hp/mana/energy/rage so hpregen doesn't add to max health (which is what it'd do now).